requirements.txt
django>=3.2.0,<=3.2.0
django-ckeditor>=6.4.0
django-modeltranslation==0.18.4
django-bootstrap-datepicker-plus==4.0.0
psycopg2==2.9.1
django-celery-results==2.4.0
django-taggit==3.0.0
validate_email==1.3
django-bootstrap-datepicker-plus==4.0.0
django-redis==5.2.0
Pillow==9.2.0